Add files via upload
This is what I have so far for trying to get the sword tip to hit the marker at (1,0,0). 

I've included pirouetteandwave with its launch file repurposed so that instead of waving it should touch the point but it doesn't. I think the main problem is the frames, as I'm doing some in world frame and some in robot frame and they're different so it's messing things up. Need help with this. 

*Note that in order to make the marker appear you have to add "Marker" in ros2 then